Minor refactor into modules & classes
