+++ app/domain/ingredient_extractor/word.rb
@@ -1,0 +1,13 @@
+module IngredientExtractor
+  class Word < SimpleDelegator
+    attr_accessor :include_in_name
+
+    def initialize(word)
+      super(word)
+      @include_in_name = true
+    end
+
+    alias :include_in_name? :include_in_name
+
+  end
+end
